+++ internal/domain/storage/gorm/scheme/gormRefreshToken.go
@@ -4,6 +4,6 @@ type RefreshToken struct {
 	ID                 uint   `gorm:"primaryKey"`
 	Token              string `gorm:"index;unique"`
 	Login              string `gorm:"index;unique"`
-	ExpirationTimeUnix int
-	CreateTimeUnix     int
+	ExpirationTimeUnix int64
+	CreateTimeUnix     int64
 }
